@@ -1,23 +1,13 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Builder;
+﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.HttpsPolicy;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
 using RankedElo.Core.Interfaces;
 using RankedElo.Core.Services;
 using RankedElo.Persistence.Contexts;
 using RankedElo.Persistence.Services;
-using Swashbuckle.AspNetCore.Swagger;
-using Microsoft.EntityFrameworkCore.Sqlite;
+using Microsoft.Extensions.Hosting;
 
 namespace RankedElo.Web
 {
@@ -34,21 +24,17 @@ namespace RankedElo.Web
         {
             services.AddScoped<IMatchService, MatchService>();
             services.AddScoped<IMatchRepository, MatchRepository>();
-            services.AddScoped<IPlayerService, PlayerService>();
+            services.AddScoped<IPlayerRepository, PlayerRepository>();
 
             services.AddDbContext<RankedEloDbContext>(options =>
                 options.UseSqlite("Data Source=rankedElo.db"));
 
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
-
-            services.AddSwaggerGen(c =>
-            {
-                c.SwaggerDoc("v1", new Info { Title = "FoosballApi", Version = "v1" });
-            });
+            services.AddControllers();
+            services.AddSwaggerGen();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
@@ -65,9 +51,11 @@ namespace RankedElo.Web
             app.UseSwagger();
             app.UseSwaggerUI(c =>
             {
-                c.SwaggerEndpoint("/swagger/v1/swagger.json", "FoosballApi V1");
+                c.SwaggerEndpoint("./swagger/v1/swagger.json", "Ranked Foosball V1");
+                c.RoutePrefix = string.Empty;
             });
-            app.UseMvc();
+            app.UseRouting();
+            app.UseEndpoints(c => c.MapControllers());
         }
     }
 }
